pyproject.toml
[project]
name = "linkages"
version = "0.1.0"
description = "LLM-powered product-to-HS-code classification pipeline"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.57.1",
    "openai>=1.96.0",
    "faiss-cpu>=1.11.0",
    "sentence-transformers>=5.0.0",
    "polars>=1.31.0",
    "pandas>=2.3.0",
    "openpyxl>=3.1.5",
    "python-dotenv>=1.0.0",
    "tqdm>=4.67.1",
]

[dependency-groups]
dev = ["ipykernel>=6.29.5"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/linkages"]

[project.scripts]
linkages-classify = "linkages.pipeline:main"
